organized tests/, renamed github actions & included dataloader test
